Update linear algebra code structure.
