+++ symctrl/matrix/eig.hpp
@@ -7,12 +7,6 @@
 
 namespace Controls {
 
-// Check if matrix is square.
-bool is_square(SymEngine::MatrixBase &A);
-
-SymEngine::RCP<const SymEngine::Basic> norm2(SymEngine::vec_basic &v);
-void normalize(SymEngine::vec_basic &v);
-
 // Should this use the Schur decomposition, or the polynomial solve method?
 void eigenvalues(SymEngine::DenseMatrix &A,
                          SymEngine::vec_basic &l,
@@ -22,23 +16,7 @@ void eigenvalues(SymEngine::DenseMatrix &A,
 void eigenvectors();
 
 // Compute the jordan form using eigenvector matrix.
-void compute_jordan_form();
-
-void hessenberg(SymEngine::DenseMatrix &A,
-                SymEngine::DenseMatrix &result);
-
-void schur(SymEngine::DenseMatrix &A,
-           SymEngine::DenseMatrix &U,
-           SymEngine::DenseMatrix &T,
-           size_t n_iter = 10);
-
-// Evaluate matrix to find reduced row echelon form.
-// Gauss-Jordan elimination with partial pivoting.
-void rref();
-// Use rref to find rank? Or some rank-revealing decomposition RRQR?
-void rank();
-
-void companion_matrix(SymEngine::vec_basic &c, SymEngine::DenseMatrix &result);
+void jordan_form();
 
 } // Controls
