+++ symctrl/matrix/house.cc
@@ -1,0 +1,17 @@
+#include <symengine/add.h>
+#include <symengine/pow.h>
+#include <symengine/visitor.h>
+
+#include "linalg.hpp"
+
+namespace Controls {
+
+// ----------------------------------------------------------------------
+// Compute Householder Reflector
+//
+void householder_reflector(size_t i, SymEngine::vec_basic &v) {
+  size_t n = v.size();
+
+}
+
+} // Controls
