+++ symctrl/matrix/qr.cc
@@ -1,0 +1,29 @@
+#include "linalg.hpp"
+
+namespace Controls {
+
+// ----------------------------------------------------------------------
+// Greedy QR
+//
+void qr_greedy(SymEngine::DenseMatrix &M,
+               SymEngine::DenseMatrix &Q,
+               SymEngine::DenseMatrix &R) {
+  SymEngine::DenseMatrix P;
+
+  qr_greedy(M, Q, R, P);
+}
+
+void qr_greedy(SymEngine::DenseMatrix &M,
+               SymEngine::DenseMatrix &Q,
+               SymEngine::DenseMatrix &R,
+               SymEngine::DenseMatrix &P) {
+  //
+  R = M;
+  Q = SymEngine::DenseMatrix()
+  SymEngine::eye(Q);
+  SymEngine::eye(P);
+
+
+}
+
+} // Controls
